requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-dotenv==1.0.0
Flask-Limiter==3.5.0